Scope and Flexibility Changes
Modified most function definitions to make them private. The state
function is the only public function now. Replaced all DBTa and DBTb
with global variable Min_DBT and Max_DBT and also added tolerance
variable TOL. This makes the module a little bit more flexibility and
easier to modified in the the future.